+++ CHANGELOG.md
@@ -9,6 +9,8 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+## [v1.16.0](https://github.com/allenai/beaker-py/releases/tag/v1.16.0) - 2023-01-26
+
 ### Added
 
 - Added missing `user_restrictions` field to `Cluster` data model.